Made Window separate module
